refactor : 안드로이드 오목에 인터페이스 적용
